Updated URLS in the urlJson object
It was throwing errors on a site that was SSL secured, so I updated the code to use https instead.  Twitter, LinkedIn and Pinterest have been tested and work and those are the ones I changed
I didn't change the urls for digg or delicious because they didn't work with the quick test I did.